requirements.txt
Flask==2.2.5
python-telegram-bot==13.15
SQLAlchemy==1.4.46
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==0.21.0